producer/pyproject.toml
[project]
name = "producer"
version = "0.0.1"
dependencies = [
    "kafka-python>=2.3.0",
    "shared",
]

[tool.uv.sources]
shared = { path = "../shared", editable = true }